temp_backend_deploy/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
requests
beautifulsoup4
gunicorn
